+++ docs/plugins/builders/component_generator.rb
@@ -37,7 +37,8 @@ class Builders::ComponentGenerator < SiteBuilder
         })
 
         path = metadata.parent_module.path
-        import_name = metadata.parent_module.exports.find { |hash| hash["name"] == "default" }["declaration"]["name"]
+        import_name = metadata.parent_module.exports.find { |hash| hash.name == "default" }.declaration.name
+
         tag_name = metadata.tagName
 
         slots = metadata.slots